enable fse v2 by default
