+++ fe/fe-core/src/test/java/com/starrocks/persist/OperationTypeTest.java
@@ -179,7 +179,7 @@ public class OperationTypeTest {
         Assertions.assertFalse(OperationType.IGNORABLE_OPERATIONS.contains(OperationType.OP_INSERT_OVERWRITE_STATE_CHANGE));
         Assertions.assertFalse(OperationType.IGNORABLE_OPERATIONS.contains(OperationType.OP_INVALID));
         Assertions.assertFalse(OperationType.IGNORABLE_OPERATIONS.contains(OperationType.OP_LEADER_INFO_CHANGE_V2));
-        Assertions.assertFalse(OperationType.IGNORABLE_OPERATIONS.contains(OperationType.OP_MODIFY_TABLE_ADD_OR_DROP_COLUMNS));
+        Assertions.assertFalse(OperationType.IGNORABLE_OPERATIONS.contains(OperationType.OP_FAST_ALTER_TABLE_COLUMNS));
         Assertions.assertFalse(OperationType.IGNORABLE_OPERATIONS.contains(OperationType.OP_RECOVER_DB_V2));
         Assertions.assertFalse(OperationType.IGNORABLE_OPERATIONS.contains(OperationType.OP_RECOVER_PARTITION_V2));
         Assertions.assertFalse(OperationType.IGNORABLE_OPERATIONS.contains(OperationType.OP_RECOVER_TABLE_V2));
